requirements.txt
gradio
Pillow
numpy
torch
transformers
requests